pyproject.toml
[tool.poetry]
name = "ss13"
version = "0.1.0"
description = ""
authors = ["Kyle Oliver <oliverke@mail.uc.edu>"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.22.3"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
